- delegate pool balance change to uint64(currency.Coin)

+++ code/go/0chain.net/smartcontract/storagesc/benchmark_setup.go
@@ -638,7 +638,7 @@ func GetMockBlobberStakePools(
 					ProviderType: int(spenum.Blobber),
 					ProviderID:   bId,
 					DelegateID:   sp.Pools[id].DelegateID,
-					Balance:      int64(sp.Pools[id].Balance),
+					Balance:      sp.Pools[id].Balance,
 					Reward:       0,
 					TotalReward:  0,
 					TotalPenalty: 0,
